+++ ggshield/cmd/hmsl/check.py
@@ -7,8 +7,6 @@ from requests import HTTPError
 from ggshield.cmd.common_options import add_common_options, json_option
 from ggshield.cmd.hmsl.decrypt import show_results
 from ggshield.cmd.hmsl.hmsl_common_options import (
-    InputType,
-    NamingStrategy,
     full_hashes_option,
     input_arg,
     input_type_option,
@@ -18,7 +16,7 @@ from ggshield.core.config import Config
 from ggshield.core.errors import UnexpectedError
 from ggshield.core.text_utils import display_info, pluralize
 from ggshield.hmsl import Secret, get_client
-from ggshield.hmsl.collection import collect, prepare
+from ggshield.hmsl.collection import InputType, NamingStrategy, collect, prepare
 
 
 logger = logging.getLogger(__name__)
